+++ templates/attraction.html
@@ -1,0 +1,42 @@
+<!DOCTYPE html>
+<html lang="zh-TW">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>HomePage</title>
+  </head>
+  <body>
+    <header class="header container">
+      <h1 class="heading__primary"></h1>
+      <nav class="header__nav">
+        <a href="" class="nav__item"></a>
+        <a href="" class="nav__item"></a>
+      </nav>
+    </header>
+    <section class="section__hero container">
+      <h2 class="heading__secondary"></h2>
+      <p class="hero__description"></p>
+      <form action="" class="form__search-keyword">
+        <input type="text" class="form__search-keyword__input" />
+        <button class="btn__search"></button>
+      </form>
+    </section>
+    <section>
+        <div class="section__attractions">
+            <ul class="list-bar">
+                <li></li>
+                <li></li>
+                <li></li>
+                <li></li>
+            </ul>
+            <ul class="list-attractions">
+                <li class="list-attractions__item">
+                    <div class="list-attractions__item__top-box">
+                    </div>
+                    <div class="list-attractions__item__bottom-box">
+                    </div>
+                </li>
+            </ul>
+    </section>
+  </body>
+</html>
